requirements.txt
# =============================================================================
# RetinAI — Dépendances
# =============================================================================

# Core ML
tensorflow>=2.15.0,<2.17
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Computer Vision
opencv-python-headless>=4.8.0

# Experiment Tracking
mlflow>=2.10.0

# Configuration
pyyaml>=6.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Optimisation (optionnel)
tensorflow-model-optimization>=0.8.0
tf2onnx>=1.16.0

# Tests
pytest>=7.4.0
pytest-cov>=4.1.0

# Qualité de code
ruff>=0.2.0
mypy>=1.8.0
